Visual Paradigm Requirements, Use Case, Student Class Diagram
This check in contains the Visual Paradigm Requirements for the project,
the Use Case, and a first stab at the Class Diagram.  Also included is
the draft of the Java code generated by Visual Paradigm for this class
diagram which is not complete at this point but serves as a good
starting point.